refactor: update app to run and respond to messages
