+++ ai/index.js
@@ -1,0 +1,12 @@
+import { OpenAI } from 'openai';
+
+// LLM system prompt
+export const DEFAULT_SYSTEM_CONTENT = `You're an assistant in a Slack workspace.
+Users in the workspace will ask you to help them write something or to think better about a specific topic.
+You'll respond to those questions in a professional way.
+When you include markdown text, convert them to Slack compatible ones.
+When a prompt has Slack's special syntax like <@USER_ID> or <#CHANNEL_ID>, you must keep them as-is in your response.`;
+
+export const openai = new OpenAI({
+  apiKey: process.env.OPENAI_API_KEY,
+});
