+++ listeners/index.js
@@ -1,11 +1,9 @@
-const actions = require('./actions/index.js');
-const events = require('./events/index.js');
-const assistant = require('./assistant/index.js');
+import * as actions from './actions/index.js';
+import * as assistant from './assistant/index.js';
+import * as events from './events/index.js';
 
 export const registerListeners = (app) => {
   actions.register(app);
   events.register(app);
   assistant.register(app);
 };
-
-module.exports = { registerListeners };
